update and rename script

+++ cleanup.js
@@ -1,19 +1,24 @@
 const path = require("path");
 const fs = require("fs");
-
+// NOTES on how to run: :
+// 1. Store outside repo since it gets deleted (e.g.: repos/md-only-script.js)
+// 2. Run - node md-only-script.js msnews-experiences
 const args = process.argv.slice(2);
-const rootDir = process.cwd();
 const targetDir = args[0] ?? "";
+const isDryRun = args[1] == "--dry-run";
+
+console.warn("Dry run only. No files will be deleted.");
+
+if (targetDir == "") {
+    console.error("Missing required param: <target-dir>. Run as 'node cleanup.js <target-dir>'")
+    process.exit(0);
+}
+
+const rootDir = process.cwd();
 const startDir = path.join(rootDir, targetDir);
 
-const dirs_to_del = new Set();
+let dirs_to_del = new Set();
 mark_dirs_to_del(startDir)
-const doc_filetypes = [
-    ".md",
-    ".png",
-    ".jpeg",
-    ".jpg"
-]
 
 function mark_dirs_to_del(dir) {
 
@@ -31,16 +36,11 @@ function mark_dirs_to_del(dir) {
                 }
             })
 
-        let no_doc_files = true;
+        let no_md = true;
 
         for (file of files) {
-            doc_filetypes.forEach(filetype => {
-                if (file.endsWith(filetype)) {
-                    no_doc_files = false;
-                }
-            })
-
-            if (!no_doc_files) {
+            if (file.endsWith(".md")) {
+                no_md = false;
                 break;
             }
         }
@@ -50,17 +50,17 @@ function mark_dirs_to_del(dir) {
             mark_dirs_to_del(path.join(dir, child_dir_nam));
         }
 
-        let no_doc_files_child_dirs = true;
+        let no_md_child_dirs = true;
 
         for (var child_dir_name of child_dirs) {
             const child_dir = path.join(dir, child_dir_name);
             if (!dirs_to_del.has(child_dir)) {
-                no_doc_files_child_dirs = false;
+                no_md_child_dirs = false;
                 break;
             }
         }
 
-        if (no_doc_files && no_doc_files_child_dirs) {
+        if (no_md && no_md_child_dirs) {
             dirs_to_del.add(dir)
 
             for (var child_dir_name of child_dirs) {
@@ -74,12 +74,18 @@ function mark_dirs_to_del(dir) {
 
 }
 
-// console.log(dirs_to_del);
-// Delete all the dirs marked for deletion
-for (var dir of dirs_to_del) {
-    fs.rmdir(dir, { recursive: true }, error => {
-        if (error) {
-            throw error;
-        }
-    })
+dirs_to_del = Array.from(dirs_to_del);
+
+if (isDryRun) {
+    console.log(`Delete count: ${dirs_to_del.length}. Following files will be deleted:`);
+    console.log(dirs_to_del);
+} else {
+    // Delete all the dirs marked for deletion
+    for (var dir of dirs_to_del) {
+        fs.rmdir(dir, { recursive: true }, error => {
+            if (error) {
+                throw error;
+            }
+        })
+    }
 }
